+++ common/SensorManager/sensorManager.c
@@ -48,7 +48,7 @@ int sensorManagerInit(void)
 
     s_tSensorManager.t_tTaskHandler.t_ptTask = sensorManagerTask;
     s_tSensorManager.t_tTaskHandler.t_ptTaskArg = NULL;
-    s_tSensorManager.t_tTaskHandler.t_ulStackSize = OS_TASK_DEFAULT_STACK_SIZE;
+    s_tSensorManager.t_tTaskHandler.t_ulStackSize = (size_t)OS_TASK_DEFAULT_STACK_SIZE;
     s_tSensorManager.t_tTaskHandler.a_iStopFlag = OS_TASK_SECURE_FLAG;
 
     return SENSOR_MANAGER_OK;
@@ -152,11 +152,11 @@ static void updateSensorData(void *p_pvParam)
         s_tSensorManager.t_tISensors[i] = rawValuesToMm(s_tSensorManager.t_tISensors[i]);
     }
 
-    s_tSensorManager.t_tFloorSensor = GetFloorSensorValue();
+    s_tSensorManager.t_tFloorSensor = (uint16_t)GetFloorSensorValue();
 
     if (idCardGetRole() == IDCARD_ROLE_EXPLO)
     {
-        map_engine_update_vision(s_tSensorManager.t_tISensors, SENSOR_MANAGER_SENSORS_COUNT);
+        map_engine_update_vision(s_tSensorManager.t_tISensors, (uint8_t)SENSOR_MANAGER_SENSORS_COUNT);
         map_engine_update_floor_sensor(s_tSensorManager.t_tFloorSensor);
     }
 }
